refactor : 디렉토리 구성 변경
fontsize double 로 변경

블루투스 권한 추가
