+++ APP/ucms/android/app/src/main/AndroidManifest.xml
@@ -6,6 +6,8 @@
    <uses-permission android:name="android.permission.VIBRATE" />
    <uses-permission android:name="android.permission.WAKE_LOCK" />
    <uses-permission android:name="android.permission.USE_FULL_SCREEN_INTENT" />
+   <uses-permission android:name="android.permission.BLUETOOTH"/>
+   <uses-permission android:name="android.permission.BLUETOOTH_ADMIN"/>
    <application
         android:label="ucms"
         android:icon="@mipmap/launcher_icon">
