+++ APP/ucms/lib/controllers/place_controller.dart
@@ -1,13 +1,13 @@
 import 'package:get/get.dart';
-import 'package:ucms/data/places/doom.dart';
-import 'package:ucms/data/places/doomfacility.dart';
-import 'package:ucms/data/places/doomroom.dart';
-import 'package:ucms/data/places/outside_facility.dart';
-import 'package:ucms/data/places/place.dart';
-import 'package:ucms/data/position.dart';
-import 'package:ucms/data/position_list.dart';
+import 'package:ucms/model/places/doom.dart';
+import 'package:ucms/model/places/doomfacility.dart';
+import 'package:ucms/model/places/doomroom.dart';
+import 'package:ucms/model/places/outside_facility.dart';
+import 'package:ucms/model/places/place.dart';
+import 'package:ucms/model/position.dart';
+import 'package:ucms/model/position_list.dart';
 import 'package:ucms/utils/convert_utf8.dart';
-import 'package:ucms/utils/place_util/place_repository.dart';
+import 'package:ucms/repository/place_repository.dart';
 
 class PlaceController extends GetxController {
   final repository = PlaceRepository();
