+++ APP/ucms/lib/model/anomaly.dart
