+++ APP/ucms/lib/pages/page_login/login_page.dart
@@ -8,16 +8,16 @@ import 'package:get_storage/get_storage.dart';
 import 'package:ucms/components/custom_buttons.dart';
 import 'package:ucms/components/custom_screen.dart';
 import 'package:ucms/components/texts.dart';
-import 'package:ucms/data/position_list.dart';
-import 'package:ucms/data/time_list.dart';
+import 'package:ucms/model/position_list.dart';
+import 'package:ucms/model/time_list.dart';
 import 'package:ucms/pages/page_cohort/cohort_main.dart';
 import 'package:ucms/pages/page_login/register_page.dart';
 import 'package:ucms/pages/page_user/user_main.dart';
 import 'package:ucms/theme/size.dart';
-import 'package:ucms/utils/cohort_util/cohort_controller.dart';
-import 'package:ucms/utils/place_util/place_controller.dart';
+import 'package:ucms/controllers/cohort_controller.dart';
+import 'package:ucms/controllers/place_controller.dart';
 import 'package:ucms/utils/snackbar.dart';
-import 'package:ucms/utils/user_util/user_controller.dart';
+import 'package:ucms/controllers/user_controller.dart';
 import 'package:ucms/utils/validate.dart';
 
 class LoginPage extends StatelessWidget {
