+++ APP/ucms/lib/providers/place_provider.dart
@@ -1,5 +1,5 @@
 import 'package:get/get.dart';
-import 'package:ucms/data/hostnames.dart';
+import 'package:ucms/model/hostnames.dart';
 
 class PlaceProvider extends GetConnect {
   //doom
