+++ APP/ucms/lib/repository/place_repository.dart
@@ -1,9 +1,9 @@
 
 import 'package:get/get.dart';
-import 'package:ucms/data/dto/server_multi_resp_dto.dart';
-import 'package:ucms/data/dto/server_resp_dto.dart';
+import 'package:ucms/model/dto/server_multi_resp_dto.dart';
+import 'package:ucms/model/dto/server_resp_dto.dart';
 import 'package:ucms/utils/convert_utf8.dart';
-import 'package:ucms/utils/place_util/place_provider.dart';
+import 'package:ucms/providers/place_provider.dart';
 
 class PlaceRepository {
   final PlaceProvider _placeProvider = PlaceProvider();
